More fixes

+++ controllers/user.controller.js
@@ -71,7 +71,7 @@ exports.user_login = (req,res)=>{
     });
 };
 
-// BEGIN LIST (Protected - only for companyUsers & company)
+// BEGIN LIST (Protected - only for companyUsers)
 exports.user_list = (req,res)=>{
     jwt.verify(req.token, 'secureKey', (err, authorizedData) => {
         if(err){
@@ -80,20 +80,27 @@ exports.user_list = (req,res)=>{
             res.end();
         }
         else {
-            if(authorizedData.companyUser.id || authorizedData.company.id) {
-                db.User.findAll({})
-                .then(data=>{
+            if(authorizedData.companyUser) {
+                if(authorizedData.companyUser.id) {
+                    db.User.findAll({})
+                    .then(data=>{
+                        res.setHeader('Content-type','application/json ; charset=utf-8');
+                        res.json(data);
+                        res.status(200);
+                        res.end();
+                    })
+                    .catch(error=>{
+                        res.setHeader('Content-type','application/json ; charset=utf-8');
+                        res.json(error);
+                        res.status(400).send('400 ERROR');
+                        res.end();
+                    });
+                }
+                else {
                     res.setHeader('Content-type','application/json ; charset=utf-8');
-                    res.json(data);
-                    res.status(200);
+                    res.sendStatus(403).send('403 - ACCESS DENIED');
                     res.end();
-                })
-                .catch(error=>{
-                    res.setHeader('Content-type','application/json ; charset=utf-8');
-                    res.json(error);
-                    res.status(400).send('400 ERROR');
-                    res.end();
-                });
+                }
             }
             else {
                 res.setHeader('Content-type','application/json ; charset=utf-8');
@@ -104,7 +111,7 @@ exports.user_list = (req,res)=>{
     });
 };
 
-// BEGIN DETAILS (Protected - only for companyUsers, company and account owner)
+// BEGIN DETAILS (Protected - only for companyUsers and account owner)
 exports.user_details = (req,res)=>{
     jwt.verify(req.token, 'secureKey', (err, authorizedData) => {
         if(err){
@@ -113,7 +120,7 @@ exports.user_details = (req,res)=>{
             res.end();
         }
         else {
-            if(authorizedData.companyUser.id || authorizedData.company.id || authorizedData.user.id==req.params.id) {
+            if(authorizedData.companyUser || authorizedData.user.id==req.params.id) {
                 db.User.findOne({
                     where:{
                         'id': req.params.id
@@ -184,39 +191,46 @@ exports.user_edit = (req,res)=>{
             res.end();
         }
         else {
-            if(authorizedData.user.id==req.params.id) {
-                bcrypt.hash(req.body.password, saltRounds, (err, hash)=> {
-                    db.User.update({
-                        about:req.body.about,
-                        available:req.body.available,
-                        email:req.body.email,
-                        facebook:req.body.facebook,
-                        firstName:req.body.firstName,
-                        lastName:req.body.lastName,
-                        linkedin:req.body.linkedin,
-                        password:hash,
-                        twitter:req.body.twitter,
-                        keyWordOne_id:req.body.keyWordOne_id,
-                        keyWordTwo_id:req.body.keyWordTwo_id,
-                        keyWordThree_id:req.body.keyWordThree_id
-                        },{
-                        where:{
-                            'id':req.params.id
-                        }
-                    })
-                    .then(data=>{
-                        res.setHeader('Content-type','application/json ; charset=utf-8');
-                        res.json(data);
-                        res.status(200);
-                        res.end();
-                    })
-                    .catch(error=>{
-                        res.setHeader('Content-type','application/json ; charset=utf-8');
-                        res.json(error);
-                        res.status(400).send('400 ERROR');
-                        res.end();
+            if(authorizedData.user) {
+                if(authorizedData.user.id==req.params.id) {
+                    bcrypt.hash(req.body.password, saltRounds, (err, hash)=> {
+                        db.User.update({
+                            about:req.body.about,
+                            available:req.body.available,
+                            email:req.body.email,
+                            facebook:req.body.facebook,
+                            firstName:req.body.firstName,
+                            lastName:req.body.lastName,
+                            linkedin:req.body.linkedin,
+                            password:hash,
+                            twitter:req.body.twitter,
+                            keyWordOne_id:req.body.keyWordOne_id,
+                            keyWordTwo_id:req.body.keyWordTwo_id,
+                            keyWordThree_id:req.body.keyWordThree_id
+                            },{
+                            where:{
+                                'id':req.params.id
+                            }
+                        })
+                        .then(data=>{
+                            res.setHeader('Content-type','application/json ; charset=utf-8');
+                            res.json(data);
+                            res.status(200);
+                            res.end();
+                        })
+                        .catch(error=>{
+                            res.setHeader('Content-type','application/json ; charset=utf-8');
+                            res.json(error);
+                            res.status(400).send('400 ERROR');
+                            res.end();
+                        });
                     });
-                });
+                }
+                else {
+                    res.setHeader('Content-type','application/json ; charset=utf-8');
+                    res.sendStatus(403).send('ERROR: ACCESS DENIED');
+                    res.end();
+                }
             }
             else {
                 res.setHeader('Content-type','application/json ; charset=utf-8');
@@ -236,26 +250,33 @@ exports.user_picture = (req,res)=>{
             res.end();
         }
         else {
-            if(authorizedData.user.id==req.params.id) {
-                db.User.update({
-                    picture:'./uploads/'+req.file.filename
-                    },{
-                    where:{
-                        'id':req.params.id
-                    }
-                })
-                .then(data=>{
+            if(authorizedData.user) {
+                if(authorizedData.user.id==req.params.id) {
+                    db.User.update({
+                        picture:'./uploads/'+req.file.filename
+                        },{
+                        where:{
+                            'id':req.params.id
+                        }
+                    })
+                    .then(data=>{
+                        res.setHeader('Content-type','application/json ; charset=utf-8');
+                        res.json('Data successfully uploaded');
+                        res.status(200);
+                        res.end();
+                    })
+                    .catch(error=>{
+                        res.setHeader('Content-type','application/json ; charset=utf-8');
+                        res.json(error);
+                        res.status(400).send('400 ERROR');
+                        res.end();
+                    });
+                }
+                else {
                     res.setHeader('Content-type','application/json ; charset=utf-8');
-                    res.json('Data sucdessfully uploaded');
-                    res.status(200);
+                    res.sendStatus(403).send('ERROR: ACCESS DENIED');
                     res.end();
-                })
-                .catch(error=>{
-                    res.setHeader('Content-type','application/json ; charset=utf-8');
-                    res.json(error);
-                    res.status(400).send('400 ERROR');
-                    res.end();
-                });
+                }
             }
             else {
                 res.setHeader('Content-type','application/json ; charset=utf-8');
@@ -266,7 +287,7 @@ exports.user_picture = (req,res)=>{
     });
 };
 
-// BEGIN CV UPLOAD (Protected - only for account owner)
+// BEGIN CV UPLOAD (Protected - only for account owner) // PROBLEM WITH SEQUELIZE (SAVING TO PICTURE INSTEAD OF CV FOR SOME REASON ?!?)
 exports.user_cv = (req,res)=>{
     jwt.verify(req.token, 'secureKey', (err, authorizedData) => {
         if(err){
@@ -275,26 +296,33 @@ exports.user_cv = (req,res)=>{
             res.end();
         }
         else {
-            if(authorizedData.user.id==req.params.id) {
-                db.User.update({
-                    cv:'./uploads/'+req.file.filename
-                    },{
-                    where:{
-                        'id':req.params.id
-                    }
-                })
-                .then(data=>{
+            if(authorizedData.user) {
+                if(authorizedData.user.id==req.params.id) {
+                    db.User.update({
+                        cv:'./uploads/'+req.file.filename
+                        },{
+                        where:{
+                            'id':req.params.id
+                        }
+                    })
+                    .then(data=>{
+                        res.setHeader('Content-type','application/json ; charset=utf-8');
+                        res.json('Data successfully uploaded');
+                        res.status(200);
+                        res.end();
+                    })
+                    .catch(error=>{
+                        res.setHeader('Content-type','application/json ; charset=utf-8');
+                        res.json(error);
+                        res.status(400).send('400 ERROR');
+                        res.end();
+                    });
+                }
+                else {
                     res.setHeader('Content-type','application/json ; charset=utf-8');
-                    res.json('Data successfully uploaded');
-                    res.status(200);
+                    res.sendStatus(403).send('ERROR: ACCESS DENIED');
                     res.end();
-                })
-                .catch(error=>{
-                    res.setHeader('Content-type','application/json ; charset=utf-8');
-                    res.json(error);
-                    res.status(400).send('400 ERROR');
-                    res.end();
-                });
+                }
             }
             else {
                 res.setHeader('Content-type','application/json ; charset=utf-8');
@@ -314,24 +342,31 @@ exports.user_delete = (req,res)=>{
             res.end();
         }
         else {
-            if(authorizedData.user.id==req.params.id) {
-                db.User.destroy({
-                    where:{
-                        'id': req.params.id
-                    }
-                })
-                .then(data=>{
+            if(authorizedData.user) {
+                if(authorizedData.user.id==req.params.id) {
+                    db.User.destroy({
+                        where:{
+                            'id': req.params.id
+                        }
+                    })
+                    .then(data=>{
+                        res.setHeader('Content-type','application/json ; charset=utf-8');
+                        res.json('Data successfully deleted');
+                        res.status(200);
+                        res.end();
+                    })
+                    .catch(error=>{
+                        res.setHeader('Content-type','application/json ; charset=utf-8');
+                        res.json(error);
+                        res.status(400).send('400 ERROR');
+                        res.end();
+                    });
+                }
+                else {
                     res.setHeader('Content-type','application/json ; charset=utf-8');
-                    res.json('Data successfully deleted');
-                    res.status(200);
+                    res.sendStatus(403).send('ERROR: ACCESS DENIED');
                     res.end();
-                })
-                .catch(error=>{
-                    res.setHeader('Content-type','application/json ; charset=utf-8');
-                    res.json(error);
-                    res.status(400).send('400 ERROR');
-                    res.end();
-                });
+                }
             }
             else {
                 res.setHeader('Content-type','application/json ; charset=utf-8');
